requirements.txt
streamlit
streamlit-option-menu
langchain
openai
python-dotenv